WIP - Moved extension node into its own file and removed the resolver helper
