+++ docs/faq.md
@@ -1,13 +1,27 @@
 # FAQ <!-- omit in toc -->
 
+- [History](#history)
+  - [Fork](#fork)
 - [Implementation](#implementation)
   - [What's the story behind Extension Node Implementation?](#whats-the-story-behind-extension-node-implementation)
 
 This documentation is meant to capture common questions that come up and act
 as a supplement or secondary reference to the primary documentation.
 
+## History
+
+### Fork
+
+This library was originally forked off of [celestiaorg/smt](https://github.com/celestiaorg/smt)
+which was archived on Feb 27th, 2023.
+
 ## Implementation
 
 ### What's the story behind Extension Node Implementation?
 
-[smt](./smt.md#extension-nodes)
+The [SMT extension node](./smt.md#extension-nodes) is very similar to that of
+Ethereum's [Modified Merkle Patricia Trie](https://ethereum.org/developers/docs/data-structures-and-encoding/patricia-merkle-trie).
+
+A quick primer on it can be found in this [5P;1R post](https://olshansky.substack.com/p/5p1r-ethereums-modified-merkle-patricia).
+
+WIP
